# Conflicts:
#	.gitignore
#	src/mc/deps/core/file/Path.h
#	src/mc/deps/core/math/Color.h
#	src/mc/deps/core/string/HashedString.h
#	src/mc/deps/core/utility/EnableNonOwnerReferences.h
#	src/mc/deps/json/Value.h
#	src/mc/network/NetworkIdentifier.h
#	src/mc/network/NetworkIdentifierWithSubId.h
#	src/mc/network/RakNetConnector.h
#	src/mc/platform/CallStack.h
#	src/mc/platform/UUID.h
#	src/mc/server/commands/CommandParameterData.h
#	src/mc/util/ExpressionNode.h
#	src/mc/world/Container.h
#	src/mc/world/actor/AreaEffectCloud.h
#	src/mc/world/actor/player/Player.h
#	src/mc/world/item/components/ThrowableItemComponent.h
#	src/mc/world/item/components/UseModifiersItemComponent.h
#	src/mc/world/level/levelgen/structure/structurepools/StructurePoolElement.h
#	xmake.lua

+++ src/mc/entity/components/DeathTickingType.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class DeathTickingType : int {
+enum class DeathTickingType : uchar {
     Mob         = 0,
     EnderDragon = 1,
 };
